+++ component/common/application/matter/core/matter_command.h
@@ -12,9 +12,9 @@ class Command
 {
 public:
     // TODO: add callback
-    Command(uint32_t _command_id, uint32_t _cluster_id, uint32_t _endpoint_id, uint16_t command_flags, callback_t command_callback);
+    Command(uint32_t _command_id, uint32_t _cluster_id, uint16_t _endpoint_id, uint16_t command_flags, callback_t command_callback);
     ~ Command();
-    uint32_t endpoint_id;
+    uint16_t endpoint_id;
     uint32_t cluster_id;
     uint32_t command_id;
     uint16_t flags;
